pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "convertkit-py"
version = "0.1.0"
authors = [
    { name = "Andres Kull", email = "andres.a.kull@gmail.com" },
]
description = "Python client for the ConvertKit API"
readme = "README.md"
requires-python = ">=3.11"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "requests>=2.32.3",
    "numpy>=1.26.4",
    "python-dotenv>=1.0.1",
]

[tool.setuptools.packages.find]
where = ["src"]
include = ["convertkit"]

[project.urls]
"Homepage" = "https://github.com/andreskull/convertkit-py"
"Bug Tracker" = "https://github.com/andreskull/convertkit-py/issues"